+++ src/App.tsx
@@ -1,49 +1,16 @@
-import { useEffect, useState } from "react";
-import { dummyData } from "./data/todos";
 import AddTodo from "./components/AddTodo";
 import TodoList from "./components/TodoList";
 import TodoSummary from "./components/TodoSummary";
-import { Todo } from "./types/todo";
+import useTodos from "./hooks/useTodos";
 
 function App() {
-  const [todos, setTodos] = useState(() => {
-    const savedTodos: Todo[] = JSON.parse(
-      localStorage.getItem("todos") || "[]",
-    );
-
-    return savedTodos.length > 0 ? savedTodos : dummyData;
-  });
-
-  useEffect(() => {
-    localStorage.setItem("todos", JSON.stringify(todos));
-  }, [todos]);
-
-  function setTodoCompleted(id: number, completed: boolean) {
-    setTodos((prevTodos) =>
-      prevTodos.map((todo) => (todo.id === id ? { ...todo, completed } : todo)),
-    );
-  }
-
-  function addTodo(title: string) {
-    // we are not able to directly modify existing todos so we create a new array
-    setTodos((prevTodos) => [
-      {
-        id: Date.now(),
-        title,
-        completed: false,
-      },
-      ...prevTodos,
-    ]);
-  }
-
-  function deleteTodo(id: number) {
-    setTodos((prevTodos) => prevTodos.filter((todo) => todo.id != id));
-  }
-
-  function deleteAllCompletedTodos() {
-    setTodos((prevTodos) => prevTodos.filter((todo) => !todo.completed));
-  }
-
+  const {
+    addTodo,
+    setTodoCompleted,
+    deleteAllCompletedTodos,
+    deleteTodo,
+    todos,
+  } = useTodos();
   return (
     <main className="py-10 bg-sky-50 h-screen overflow-y-auto">
       <h1 className="font-bold text-3xl text-center space-y-5 p-5">
